NSOF-5982 metaport_cluster: support data-source name param
Even though name is not unique for metaport clusters - costumer's feedback
brought the need to onboard metaports and metaport clusters outside of tf code and later - use
the metaport name to find it and attach mapped elements to it from
terraform.

Because it's not unique - in case there is more then one metaport cluster with the
specified name - an error will be returned.

Customer's ticket: https://github.com/nsofnetworks/terraform-provider-pfptmeta/issues/11