NSOF-9189 idp: remove PEM validation when creating SAML IdP resources
The PEM validation function used is meant to validate public keys, and
indeed the unittest does not use a full certificate.

There IdP resource requires a full certificate, so this validation isn't
relevant, and prevents the provisioning of proper SAML resources.

Furthermore there's not much point in having the terraform provider
validate the certificates to this extent as the API does its own validation.

Remove the validation call.